* Created passthrough factory to allocate shim proxy from detected type * Created model passthrough shim interface. * Proxy ap214 not created yet (effects of passthrough not being implemented are roughly the same as a nil implementation anyway). * Created IFC shim proxy. * IfcApi ported over to use proxy passthrough. * Tests pass.
