+++ package.json
@@ -100,6 +100,5 @@
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4"
   },
-  "main": "./compiled/src/index.js",
-  "types": "./compiled/src/index.d.ts"
+  "type": "module"
 }
