+++ src/shim/ifc_api.ts
@@ -1,23 +1,11 @@
-import IfcStepParser from '../ifc/ifc_step_parser'
-import ParsingBuffer from '../parsing/parsing_buffer'
-import { ParseResult } from '../step/parsing/step_parser'
-import { EntityTypesIfcCount } from '../ifc/ifc4_gen/entity_types_ifc.gen'
-import IfcStepModel from '../ifc/ifc_step_model'
-import { IfcGeometryExtraction } from '../ifc/ifc_geometry_extraction'
-import { BlendMode, ConwayGeometry, GeometryObject }
+import { ConwayGeometry }
   from '../../dependencies/conway-geom/conway_geometry'
-import { CanonicalMeshType } from '../core/canonical_mesh'
-import { CanonicalMaterial } from '../core/canonical_material'
-import { shimIfcEntityMap, shimIfcEntityReverseMap } from './shim_schema_mapping'
-import { IfcSceneBuilder } from '../ifc/ifc_scene_builder'
 import * as glmatrix from 'gl-matrix'
-import { Properties } from './properties'
-import { FromRawLineData } from './ifc2x4_helper'
-import { ExtractResult } from '../core/shared_constants'
 import { versionString } from '../version/version'
 import Logger from '../logging/logger'
 import Environment from '../utilities/environment'
-import Memory from '../memory/memory'
+import { IfcApiModelPassthrough } from './ifc_api_model_passthrough'
+import { IfcApiModelPassthroughFactory } from './ifc_api_model_passthrough_factory'
 
 
 export * from './ifc2x4'
@@ -106,18 +94,7 @@ export class IfcAPI {
   isWasmPathAbsolute = false
   settings: Loadersettings | undefined
   globalModelIDCounter = 0
-  models: Map<number,
-    [IfcStepModel,
-      IfcSceneBuilder,
-      Map<number, [Vector<PlacedGeometry>, FlatMesh]>,
-      Map<number, [GeometryObject, CanonicalMaterial, number[]]>,
-      Vector<FlatMesh>, glmatrix.mat4]> =
-      new Map<number,
-      [IfcStepModel,
-        IfcSceneBuilder,
-        Map<number, [Vector<PlacedGeometry>, FlatMesh]>,
-        Map<number, [GeometryObject, CanonicalMaterial, number[]]>,
-        Vector<FlatMesh>, glmatrix.mat4]>()
+  models = new Map<number, IfcApiModelPassthrough>()
   conwaywasm = new ConwayGeometry()
   _isCoordinated: boolean = false
   linearScalingFactor: number = 1
@@ -130,11 +107,6 @@ export class IfcAPI {
       0, 1, 0, 0,  // Third column
       0, 0, 0, 1,   // Fourth column
   )
-
-  /**
-   * Contains all the logic and methods regarding properties, psets, qsets, etc.
-   */
-  properties = new Properties(this)
 
   /**
    * Initializes the WASM module (WebIFCWasm), required before using any other functionality.
@@ -180,238 +152,27 @@ export class IfcAPI {
    * @return {number} model ID
    */
   OpenModel(data: Uint8Array, settings?: Loadersettings): number {
-    const allTimeStart = Date.now()
-    const parser = IfcStepParser.Instance
-    const bufferInput = new ParsingBuffer(data)
-    const [stepHeader, result0] = parser.parseHeader(bufferInput)
-    const tempModelID = this.globalModelIDCounter
-    Logger.createStatistics(tempModelID)
 
-    const statistics = Logger.getStatistics(tempModelID)
+    const modelIdResult = this.globalModelIDCounter
 
+    const result =
+      IfcApiModelPassthroughFactory.from(
+          modelIdResult,
+          data,
+          this.wasmModule,
+          settings)
 
-    switch (result0) {
-      case ParseResult.COMPLETE:
-
-        break
-
-      case ParseResult.INCOMPLETE:
-
-        Logger.warning('Parse incomplete but no errors')
-        statistics?.setLoadStatus('HEADER PARSE: INCOMPLETE')
-        break
-
-      case ParseResult.INVALID_STEP:
-
-        Logger.error('Error: Invalid STEP detected in parse, but no syntax error detected')
-        statistics?.setLoadStatus('HEADER PARSE: INVALID_STEP')
-        break
-
-      case ParseResult.MISSING_TYPE:
-
-        Logger.warning('Error: missing STEP type, but no syntax error detected')
-        statistics?.setLoadStatus('HEADER PARSE: MISSING_TYPE')
-        break
-
-      case ParseResult.SYNTAX_ERROR:
-
-        Logger.error(`Error: Syntax error detected on line ${bufferInput.lineCount}`)
-        statistics?.setLoadStatus('HEADER PARSE: SYNTAX_ERROR')
-        break
-
-      default:
-    }
-
-
-    const parseStartTime = Date.now()
-    const model = parser.parseDataToModel(bufferInput)[1]
-    const parseEndTime = Date.now()
-
-    if (model === void 0) {
-      Logger.error('[OpenModel]: model === undefined')
-      statistics?.setLoadStatus('PARSE_FAIL')
+    if ( result === void 0 ) {
       return -1
     }
 
-    statistics?.setParseTime(parseEndTime - parseStartTime)
+    this.globalModelIDCounter++
 
-    // TODO(nickcastel50): Doing geometry extraction in here for now...
-    // parse + extract data model + geometry data
-    const conwayGeometry = new IfcGeometryExtraction(this.conwaywasm, model)
+    this.models.set( modelIdResult, result )
 
-    const startTime = Date.now()
-    const [extractionResult, scene] =
-      conwayGeometry.extractIFCGeometryData()
-
-    const endTime = Date.now()
-    const executionTimeInMs = endTime - startTime
-
-    if (extractionResult !== ExtractResult.COMPLETE) {
-      Logger.error('[OpenModel]: Error extracting geometry, exiting...')
-      statistics?.setLoadStatus('FAIL')
-      return -1
-    }
-
-    // get linear scaling factor
-    this.linearScalingFactor = conwayGeometry.getLinearScalingFactor()
-
-    const ifcProjectName = conwayGeometry.getIfcProjectName()
-
-    if (ifcProjectName !== null) {
-      statistics?.setProjectName(ifcProjectName)
-    }
-
-    // build packed mesh model
-    // const packedMeshModel = scene.buildPackedMeshModel()
-
-    const vectorGeometryMap = new Map<number, [Vector<PlacedGeometry>, FlatMesh]>()
-
-    const geometryMap = new Map<number, [GeometryObject, CanonicalMaterial, number[]]>()
-
-    // dummy vars
-    const dummyColor = {
-      x: 0,
-      y: 0,
-      z: 0,
-      w: 0,
-    }
-
-    // Single PlacedGeometry variable
-    const singlePlacedGeometry: PlacedGeometry = {
-      color: dummyColor,
-      geometryExpressID: 0, // replace with actual ID
-      flatTransformation: this.identity,
-    }
-
-    // eslint-disable-next-line no-array-constructor
-    const placedGeometryArray = new Array<PlacedGeometry>()
-
-    // Vector of PlacedGeometry
-    const vectorOfPlacedGeometry: Vector<PlacedGeometry> = {
-      get(index: number): PlacedGeometry {
-        if (index >= placedGeometryArray.length) {
-          return singlePlacedGeometry
-        }
-
-        return placedGeometryArray[index]
-      },
-      size(): number {
-        return placedGeometryArray.length
-      },
-      push(parameter: PlacedGeometry): void {
-        placedGeometryArray.push(parameter)
-      },
-    }
-
-    // eslint-disable-next-line no-array-constructor
-    const flatMeshArray = new Array<FlatMesh>()
-    const flatMeshDummy: FlatMesh = {
-      geometries: vectorOfPlacedGeometry,
-      expressID: 0, // replace with actual expressID
-    }
-
-    // Vector of FlatMesh
-    const vectorFlatMesh: Vector<FlatMesh> = {
-      get(index: number): FlatMesh {
-        if (index >= placedGeometryArray.length) {
-          return flatMeshDummy
-        }
-
-        return flatMeshArray[index]
-      },
-      size(): number {
-        // Your implementation here
-        return flatMeshArray.length
-      },
-      push(parameter: FlatMesh): void {
-        flatMeshArray.push(parameter)
-      },
-    }
-
-    const coordinationMatrix: glmatrix.mat4 = glmatrix.mat4.create()
-
-
-    this.models.set(this.globalModelIDCounter++,
-        [model, scene, vectorGeometryMap, geometryMap, vectorFlatMesh, coordinationMatrix])
-
-    // save settings
-    this.settings = settings
-
-    let FILE_NAME = stepHeader.headers.get('FILE_NAME')
-
-    if (FILE_NAME !== void 0) {
-      // strip start / end parenthesis
-      FILE_NAME = FILE_NAME.substring(1, FILE_NAME.length - 1)
-    }
-
-    const ifcVersion = stepHeader.headers.get('FILE_SCHEMA')
-
-    const allTimeEnd = Date.now()
-
-    const allTime = allTimeEnd - allTimeStart
-
-    statistics?.setLoadStatus('OK')
-    statistics?.setTotalTime(allTime)
-
-    if (ifcVersion !== void 0) {
-      statistics?.setVersion(ifcVersion)
-    }
-
-    if (FILE_NAME !== void 0) {
-      const fileNameSplit: string[] = this.parseFileHeader(FILE_NAME)
-
-      // eslint-disable-next-line no-magic-numbers
-      if (fileNameSplit.length > 5) {
-        const preprocessorVersion = fileNameSplit[5]
-        const originatingSystem = fileNameSplit[6]
-
-        statistics?.setPreprocessorVersion(preprocessorVersion)
-        statistics?.setOriginatingSystem(originatingSystem)
-      }
-    }
-
-    statistics?.setMemoryStatistics(Memory.checkMemoryUsage())
-
-    statistics?.setGeometryTime(executionTimeInMs)
-    // eslint-disable-next-line no-magic-numbers
-    statistics?.setGeometryMemory(scene.model.geometry.calculateGeometrySize() / (1024 * 1024))
-
-
-    return tempModelID
+    return modelIdResult
   }
 
-  /**
-   *
-   * @param input - FILE_HEADER from step header
-   * @return {string[]} array of fields in FILE_NAME
-   */
-  parseFileHeader(input: string): string[] {
-    const result: string[] = []
-    let currentSegment = ''
-    let parenthesesCount = 0
-
-    for (const char of input) {
-      if (char === '(') {
-        parenthesesCount++
-      } else if (char === ')') {
-        parenthesesCount--
-      }
-
-      if (char === ',' && parenthesesCount === 0) {
-        result.push(currentSegment.trim())
-        currentSegment = ''
-      } else {
-        currentSegment += char
-      }
-    }
-
-    // Add the last segment if it's not empty
-    if (currentSegment.trim() !== '') {
-      result.push(currentSegment.trim())
-    }
-
-    return result
-  }
 
   /**
    * Creates a new model and returns a modelID number (unimplemented)
@@ -448,30 +209,18 @@ export class IfcAPI {
     const result = this.models.get(modelID)
 
     if (result !== void 0) {
-      // eslint-disable-next-line no-unused-vars
-      const [model, scene, placedGeometryVec, geometryMap] = result
 
-      const mapResult = geometryMap.get(geometryExpressID)
+      return result.getGeometry(geometryExpressID)
 
-      if (mapResult !== undefined) {
-
-        // eslint-disable-next-line no-unused-vars
-        const [geometryObject, _] = mapResult
-        if (geometryObject !== void 0) {
-          const clone = geometryObject.clone()
-
-          return clone
-        } else {
-          Logger.error(`[GetGeometry]: Geometry Object not found for expressID: 
-          ${geometryExpressID}`)
-        }
-      }
     } else {
+
       Logger.error('[GetGeometry]: model === undefined')
     }
 
     Logger.error('[GetGeometry]: Error - returning dummyGeometry object')
+
     const dummyGeometry: IfcGeometry = (new (this.wasmModule.IfcGeometry)())
+
     return dummyGeometry
   }
 
@@ -484,21 +233,15 @@ export class IfcAPI {
    */
   GetLine(modelID: number, expressID: number, flatten: boolean = false) {
 
-    // eslint-disable-next-line new-cap
-    const rawLineData = this.GetRawLineData(modelID, expressID)
+    const result = this.models.get(modelID)
 
-    if (rawLineData.type === -1) {
-      Logger.warning(`RawLineData null, expressID: ${expressID}`)
+    if (result === void 0) {
+
+      Logger.error('[GetLine]: model === undefined')
       return
     }
 
-    const lineData = FromRawLineData[rawLineData.type](rawLineData)
-    if (flatten) {
-      // eslint-disable-next-line new-cap
-      this.FlattenLine(modelID, lineData)
-    }
-
-    return lineData
+    return result.getLine(expressID, flatten)
   }
 
   /**
@@ -539,23 +282,18 @@ export class IfcAPI {
    *
    * @param modelID
    * @param line
+   * @return {string | undefined}
    */
   FlattenLine(modelID: number, line: any) {
-    Logger.warning('[FlattenLine]: Shim - implemented')
-    Object.keys(line).forEach((propertyName) => {
-      const property = line[propertyName]
-      // eslint-disable-next-line no-magic-numbers
-      if (property && property.type === 5) {
-        // eslint-disable-next-line new-cap
-        line[propertyName] = this.GetLine(modelID, property.value, true)
-        // eslint-disable-next-line no-magic-numbers
-      } else if (Array.isArray(property) && property.length > 0 && property[0].type === 5) {
-        for (let i = 0; i < property.length; i++) {
-          // eslint-disable-next-line new-cap
-          line[propertyName][i] = this.GetLine(modelID, property[i].value, true)
-        }
-      }
-    })
+    const result = this.models.get(modelID)
+
+    if (result === void 0) {
+
+      Logger.error('[FlattenLine]: model === undefined')
+      return
+    }
+
+    return result.flattenLine(line)
   }
 
   /**
@@ -565,120 +303,22 @@ export class IfcAPI {
    * @return {RawLineData}
    */
   GetRawLineData(modelID: number, expressID: number): RawLineData {
+
     const result = this.models.get(modelID)
 
-    if (result !== undefined) {
-      // eslint-disable-next-line no-unused-vars
-      const [model, scene] = result
+    if (result === void 0) {
 
-      const element = model.getElementByExpressID(expressID)
+      Logger.error('[GetRawLineData]: model === undefined')
 
-      const args: any[] = []
-
-      if (element !== void 0) {
-        const lineArguments = element.extractLineArguments()
-
-        const parsingBuffer = new ParsingBuffer(lineArguments)
-        if (element.expressID !== void 0) {
-          const result_ = IfcStepParser.Instance.extractArguments(parsingBuffer, element.expressID)
-          if (result_[1] === ParseResult.COMPLETE) {
-            const rawLineData: RawLineData = {
-              ID: expressID,
-              type: shimIfcEntityReverseMap[element.type],
-              arguments: result_[0],
-            }
-
-            return rawLineData
-          }
-        } else {
-          Logger.warning('element express ID null')
-        }
-
-        const rawLineData: RawLineData = {
-          ID: expressID,
-          type: shimIfcEntityReverseMap[element.type],
-          arguments: args,
-        }
-
-        return rawLineData
-      } else {
-        Logger.warning(`element === undefined, expressID: ${expressID}`)
+      return {
+        ID: expressID,
+        type: -1,
+        arguments: ['invalid'],
       }
     }
 
-    const dummyRawLineData: RawLineData = {
-      ID: expressID,
-      type: -1,
-      arguments: ['invalid'],
-    }
+    return result.getRawLineData(expressID)
 
-    return dummyRawLineData
-  }
-
-  /**
-   *
-   * @param modelID
-   * @param data
-   */
-  WriteRawLineData(modelID: number, data: RawLineData) {
-    Logger.warning('[WriteRawLineData]: Shim - Unimplemented')
-  }
-
-  /**
-   *
-   * @param modelID
-   * @param type
-   * @return {Vector<number>}
-   */
-  GetLineIDsWithType(modelID: number, type: number): Vector<number> {
-    const vectorArray: Array<number> = []
-    const expressIDVector: Vector<number> = {
-      get(index: number): number {
-        // Your implementation here
-        if (index >= vectorArray.length) {
-          return -1
-        }
-
-        return vectorArray[index]
-      },
-      size(): number {
-        // Your implementation here
-        return vectorArray.length
-      },
-
-      push(parameter: number): void {
-        vectorArray.push(parameter)
-      },
-    }
-
-    const result = this.models.get(modelID)
-    if (result !== undefined) {
-      // eslint-disable-next-line no-unused-vars
-      const [model, _] = result
-      if (type in shimIfcEntityMap) {
-        const value = shimIfcEntityMap[type]
-        // Do something with value
-        const results = model.typeIDs(value)
-        const arr = Array.from(results)
-
-        for (let arrIndex = 0; arrIndex < arr.length; ++arrIndex) {
-
-          if (arr[arrIndex].expressID !== void 0) {
-            expressIDVector.push(arr[arrIndex].expressID!)
-          } else {
-            Logger.warning('[GetLineIDsWithType] No express ID found?')
-          }
-        }
-
-      } else {
-        // Handle case where key does not exist
-        Logger.warning(`[GetLineIDsWithType] Type: ${type} does not exist in shimIfcEntityMap`)
-      }
-    } else {
-      Logger.error('[GetLineIDsWithType]: model is undefined...')
-    }
-
-    return expressIDVector
   }
 
   /**
@@ -686,52 +326,36 @@ export class IfcAPI {
    * @param modelID
    * @return {Vector<number>}
    */
-  GetAllLines(modelID: Number): Vector<number> {
-    const vectorArray: Array<number> = []
-    const expressIDVector: Vector<number> = {
-      get(index: number): number {
-        // Your implementation here
-        if (index >= vectorArray.length) {
-          return -1
-        }
+  GetAllLines(modelID: number): Vector<number> {
 
-        return vectorArray[index]
-      },
-      size(): number {
-        // Your implementation here
-        return vectorArray.length
-      },
+    const result = this.models.get(modelID)
 
-      push(parameter: number): void {
-        vectorArray.push(parameter)
-      },
-    }
+    if (result === void 0) {
 
-    const result = this.models.get(modelID as number)
+      Logger.error('[GetRawLineData]: model === undefined')
 
-    if (result !== undefined) {
-      // eslint-disable-next-line no-unused-vars
-      const [model, scene] = result
-      // TODO(nickcastel50): This is absolutely horrid but I don't know a better way yet.
-      // This implementation also kills our lazy loading...
-      for (let typeIndex = 0; typeIndex < EntityTypesIfcCount; ++typeIndex) {
-        const results = model.typeIDs(typeIndex)
-        const arr = Array.from(results)
-
-        for (let arrIndex = 0; arrIndex < arr.length; ++arrIndex) {
-
-          if (arr[arrIndex].expressID !== void 0) {
-            expressIDVector.push(arr[arrIndex].expressID!)
-          } else {
-            Logger.warning('[GetLineIDsWithType] No express ID found?')
+      const vectorArray: Array<number> = []
+      return {
+        get(index: number): number {
+          // Your implementation here
+          if (index >= vectorArray.length) {
+            return -1
           }
-        }
+
+          return vectorArray[index]
+        },
+        size(): number {
+          // Your implementation here
+          return vectorArray.length
+        },
+
+        push(parameter: number): void {
+          vectorArray.push(parameter)
+        },
       }
-    } else {
-      Logger.error('[GetLineIDsWithType]: model is undefined...')
     }
 
-    return expressIDVector
+    return result.getAllLines()
   }
 
   /**
@@ -755,21 +379,13 @@ export class IfcAPI {
    * @return {Array<number>}
    */
   GetCoordinationMatrix(modelID: number): Array<number> {
-    // TODO: Add coordination matrix to models map
+
     const result = this.models.get(modelID)
 
     if (result !== void 0) {
-      /* eslint-disable no-unused-vars */
-      const [model,
-        scene,
-        meshMap,
-        geometryMaterialTransformMap,
-        vectorFlatMesh, coordinationMatrix] = result
-      /* eslint-enable no-unused-vars */
 
-      return Array.from(coordinationMatrix)
+      return result.getCoordinationMatrix()
     }
-
 
     const coordinationMatrix: glmatrix.mat4 = glmatrix.mat4.create()
 
@@ -823,228 +439,12 @@ export class IfcAPI {
    * @param modelID
    * @param meshCallback
    */
-  StreamAllMeshes(modelID: number, meshCallback: (mesh: FlatMesh) => void) {
+  StreamAllMeshes(modelID: number, meshCallback: (mesh: FlatMesh) => void): void {
     const result = this.models.get(modelID)
 
     if (result !== void 0) {
-      const [model,
-        scene,
-        meshMap,
-        geometryMaterialTransformMap,
-        vectorFlatMesh] = result
 
-      let coordinationMatrix = result[5]
-
-      // eslint-disable-next-line no-unused-vars
-      for (const [_, nativeTransform, geometry, material, entity] of scene.walk()) {
-
-        if (geometry.type === CanonicalMeshType.BUFFER_GEOMETRY && !geometry.temporary) {
-          let material_: CanonicalMaterial | undefined
-          if (material === void 0) {
-            material_ = {
-              name: '',
-              // eslint-disable-next-line no-magic-numbers
-              baseColor: [0.8, 0.8, 0.8, 1],
-              // eslint-disable-next-line no-magic-numbers
-              legacyColor: [0.8, 0.8, 0.8, 1],
-              doubleSided: true,
-              blend: BlendMode.OPAQUE,
-            }
-          } else {
-            material_ = material
-          }
-          // extract min
-          const geomMin: glmatrix.vec3 = glmatrix.vec3.create()
-
-          const minPt = geometry.geometry.getMin()
-          geomMin[0] = minPt.x
-          geomMin[1] = minPt.y
-          geomMin[2] = minPt.z
-
-          // Create a translation matrix from geom.min
-          const translationMatrixGeomMin: glmatrix.mat4 = glmatrix.mat4.create()
-          glmatrix.mat4.fromTranslation(translationMatrixGeomMin, geomMin)
-
-          // create PlacedGeometry
-          const expressID = model.getElementByLocalID(geometry.localID)?.expressID as number
-
-          const geometryTransform = nativeTransform?.getValues()
-          let newMatrix: glmatrix.mat4 | undefined
-          if (geometryTransform !== void 0) {
-            newMatrix = glmatrix.mat4.fromValues(
-                geometryTransform[0],
-                geometryTransform[1],
-                geometryTransform[2],
-                geometryTransform[3],
-                geometryTransform[4],
-                geometryTransform[5],
-                geometryTransform[6],
-                geometryTransform[7],
-                geometryTransform[8],
-                geometryTransform[9],
-                geometryTransform[10],
-                geometryTransform[11],
-                geometryTransform[12],
-                geometryTransform[13],
-                geometryTransform[14],
-                geometryTransform[15],
-            )
-          } else {
-            // set to identity if no transform found
-            newMatrix = glmatrix.mat4.create()
-          }
-
-          if (!this._isCoordinated && this.settings?.COORDINATE_TO_ORIGIN) {
-            // coordinate the geometry to the origin
-            // eslint-disable-next-line new-cap
-            const nativePt = geometry.geometry.GetPoint(0)
-            const pt: number[] = [nativePt.x, nativePt.y, nativePt.z]
-
-            // Transform the point by the matrix.
-            const transformedPt: glmatrix.vec4 = glmatrix.vec4.create()
-            glmatrix.vec4.transformMat4(transformedPt, [pt[0], pt[1], pt[2], 1], newMatrix!)
-
-            // Create the translation matrix.
-            coordinationMatrix = glmatrix.mat4.create()
-
-            glmatrix.mat4.fromTranslation(coordinationMatrix,
-                [-transformedPt[0], -transformedPt[1], -transformedPt[2]])
-
-            const scaleMatrix = glmatrix.mat4.create()
-
-            // Create a 3D vector for scaling factors
-            const scaleVec = glmatrix.vec3.fromValues(this.linearScalingFactor,
-                this.linearScalingFactor,
-                this.linearScalingFactor)
-
-            // Scale the matrix
-            glmatrix.mat4.scale(scaleMatrix, scaleMatrix, scaleVec)
-
-            glmatrix.mat4.multiply(coordinationMatrix,
-                this.NormalizeMat,
-                coordinationMatrix)
-            glmatrix.mat4.multiply(coordinationMatrix,
-                scaleMatrix,
-                coordinationMatrix)
-
-            this._isCoordinated = true
-          }
-
-          // normalize geometry
-          if (!geometry.geometry.normalized) {
-            // eslint-disable-next-line new-cap
-            geometry.geometry.NormalizeInPlace()
-          }
-
-
-          // extract color
-          const newTransform = glmatrix.mat4.create()
-
-          // Create a 4x4 identity matrix
-          const scaleMatrix = glmatrix.mat4.create()
-
-          // Create a 3D vector for scaling factors
-          const scaleVec = glmatrix.vec3.fromValues(this.linearScalingFactor,
-              this.linearScalingFactor,
-              this.linearScalingFactor)
-
-          // Scale the matrix
-          glmatrix.mat4.scale(scaleMatrix, scaleMatrix, scaleVec)
-
-          // Perform the matrix multiplications
-          if (newMatrix !== void 0) {
-            glmatrix.mat4.multiply(newTransform, coordinationMatrix, newMatrix)
-            glmatrix.mat4.multiply(newTransform, newTransform, translationMatrixGeomMin)
-          } else {
-            glmatrix.mat4.multiply(newTransform, coordinationMatrix, newTransform)
-            glmatrix.mat4.multiply(newTransform, newTransform, translationMatrixGeomMin)
-          }
-          const newTransformArr = Array.from(newTransform)
-          geometryMaterialTransformMap.set(expressID,
-              [geometry.geometry, material_, newTransformArr])
-
-          if (entity?.localID !== void 0) {
-            if (entity?.expressID !== void 0) {
-              const mesh = meshMap.get(entity.localID)
-              if (mesh !== void 0) {
-                // set color
-                const color = {
-                  x: material_!.legacyColor[0],
-                  y: material_!.legacyColor[1],
-                  z: material_!.legacyColor[2],
-                  w: material_!.legacyColor[3],
-                }
-
-                // Single PlacedGeometry variable
-                const singlePlacedGeometry: PlacedGeometry = {
-                  color: color,
-                  geometryExpressID: expressID,
-                  flatTransformation: newTransformArr,
-                }
-
-                mesh[0].push(singlePlacedGeometry)
-                mesh[1].geometries = mesh[0]
-
-                meshMap.set(entity.localID, [mesh[0], mesh[1]])
-
-
-              } else {
-                // set color
-                const color = {
-                  x: material_!.legacyColor[0],
-                  y: material_!.legacyColor[1],
-                  z: material_!.legacyColor[2],
-                  w: material_!.legacyColor[3],
-                }
-
-                // Single PlacedGeometry variable
-                const singlePlacedGeometry_: PlacedGeometry = {
-                  color: color,
-                  geometryExpressID: expressID,
-                  flatTransformation: newTransformArr,
-                }
-
-                // eslint-disable-next-line no-array-constructor
-                const placedGeometryArray_ = new Array<PlacedGeometry>()
-
-                // Vector of PlacedGeometry
-                const vectorOfPlacedGeometry_: Vector<PlacedGeometry> = {
-                  get(index: number): PlacedGeometry {
-                    if (index >= placedGeometryArray_.length) {
-                      return singlePlacedGeometry_
-                    }
-
-                    return placedGeometryArray_[index]
-                  },
-                  size(): number {
-                    return placedGeometryArray_.length
-                  },
-                  push(parameter: PlacedGeometry): void {
-                    placedGeometryArray_.push(parameter)
-                  },
-                }
-
-                vectorOfPlacedGeometry_.push(singlePlacedGeometry_)
-
-                const singleFlatMesh: FlatMesh = {
-                  geometries: vectorOfPlacedGeometry_,
-                  expressID: entity.expressID,
-                }
-
-                meshMap.set(entity.localID, [vectorOfPlacedGeometry_, singleFlatMesh])
-              }
-            }
-          }
-        }
-      }
-
-
-      meshMap.forEach((mesh, productLocalID) => {
-
-        vectorFlatMesh.push(mesh[1])
-
-        meshCallback(mesh[1])
-      })
+      result.streamAllMeshes(meshCallback)
     }
 
     Logger.displayLogs()
@@ -1058,248 +458,14 @@ export class IfcAPI {
    * @param types
    * @param meshCallback
    */
-  StreamAllMeshesWithTypes(modelID: number,
+  streamAllMeshesWithTypes(modelID: number,
       types: Array<number>,
-      meshCallback: (mesh: FlatMesh) => void) {
+      meshCallback: (mesh: FlatMesh) => void): void {
     const result = this.models.get(modelID)
 
     if (result !== void 0) {
-      const [model,
-        scene,
-        meshMap,
-        geometryMaterialTransformMap,
-        vectorFlatMesh] = result
 
-      let coordinationMatrix = result[5]
-
-      const conwayTypesArray: number[] = []
-      types.forEach((type) => {
-        const value = shimIfcEntityMap[type]
-        // Do something with value
-        conwayTypesArray.push(value)
-      })
-
-      // eslint-disable-next-line no-unused-vars
-      for (const [_, nativeTransform, geometry, material, entity] of scene.walk()) {
-
-        if (geometry.type === CanonicalMeshType.BUFFER_GEOMETRY && !geometry.temporary) {
-
-          let material_: CanonicalMaterial | undefined
-          if (material === void 0) {
-            material_ = {
-              name: '',
-              // eslint-disable-next-line no-magic-numbers
-              baseColor: [0.8, 0.8, 0.8, 1],
-              // eslint-disable-next-line no-magic-numbers
-              legacyColor: [0.8, 0.8, 0.8, 1],
-              doubleSided: true,
-              blend: BlendMode.OPAQUE,
-            }
-          } else {
-            material_ = material
-          }
-
-          // type check
-          const typedElement = model.getElementByLocalID(geometry.localID)
-
-          if (typedElement !== void 0) {
-            if (conwayTypesArray.indexOf(typedElement.type.valueOf()) === -1) {
-              continue
-            }
-          }
-
-          // extract min
-          const geomMin: glmatrix.vec3 = glmatrix.vec3.create()
-          const minPt = geometry.geometry.getMin()
-          geomMin[0] = minPt.x
-          geomMin[1] = minPt.y
-          geomMin[2] = minPt.z
-
-          // Create a translation matrix from geom.min
-          const translationMatrixGeomMin: glmatrix.mat4 = glmatrix.mat4.create()
-          glmatrix.mat4.fromTranslation(translationMatrixGeomMin, geomMin)
-
-          // create PlacedGeometry
-          const expressID = model.getElementByLocalID(geometry.localID)?.expressID as number
-
-          const geometryTransform = nativeTransform?.getValues()
-          let newMatrix: glmatrix.mat4 | undefined
-          if (geometryTransform !== void 0) {
-            newMatrix = glmatrix.mat4.fromValues(
-                geometryTransform[0],
-                geometryTransform[1],
-                geometryTransform[2],
-                geometryTransform[3],
-                geometryTransform[4],
-                geometryTransform[5],
-                geometryTransform[6],
-                geometryTransform[7],
-                geometryTransform[8],
-                geometryTransform[9],
-                geometryTransform[10],
-                geometryTransform[11],
-                geometryTransform[12],
-                geometryTransform[13],
-                geometryTransform[14],
-                geometryTransform[15],
-            )
-          }
-
-          if (!this._isCoordinated && this.settings?.COORDINATE_TO_ORIGIN) {
-            // coordinate the geometry to the origin
-            // Assuming geom.GetPoint(0) returns a glm::dvec3, i.e., a 3D vector.
-            // In TypeScript, you can represent it as number[] or Float64Array.
-            Logger.info('Setting up coordinationMatrix')
-
-            // eslint-disable-next-line new-cap
-            const nativePt = geometry.geometry.GetPoint(0)
-            const pt: number[] = [nativePt.x, nativePt.y, nativePt.z]
-
-            // Transform the point by the matrix.
-            const transformedPt: glmatrix.vec4 = glmatrix.vec4.create()
-            glmatrix.vec4.transformMat4(transformedPt, [pt[0], pt[1], pt[2], 1], newMatrix!)
-
-            // Create the translation matrix.
-            coordinationMatrix = glmatrix.mat4.create()
-
-            glmatrix.mat4.fromTranslation(coordinationMatrix,
-                [-transformedPt[0], -transformedPt[1], -transformedPt[2]])
-
-            const scaleMatrix = glmatrix.mat4.create()
-
-            // Create a 3D vector for scaling factors
-            const scaleVec = glmatrix.vec3.fromValues(this.linearScalingFactor,
-                this.linearScalingFactor,
-                this.linearScalingFactor)
-
-            // Scale the matrix
-            glmatrix.mat4.scale(scaleMatrix, scaleMatrix, scaleVec)
-
-            glmatrix.mat4.multiply(coordinationMatrix,
-                this.NormalizeMat,
-                coordinationMatrix)
-            glmatrix.mat4.multiply(coordinationMatrix,
-                scaleMatrix,
-                coordinationMatrix)
-
-            this._isCoordinated = true
-          }
-
-          // normalize geometry
-          if (!geometry.geometry.normalized) {
-            // eslint-disable-next-line new-cap
-            geometry.geometry.NormalizeInPlace()
-          }
-
-
-          // extract color
-          const newTransform = glmatrix.mat4.create()
-
-          // Create a 4x4 identity matrix
-          const scaleMatrix = glmatrix.mat4.create()
-
-          // Create a 3D vector for scaling factors
-          const scaleVec = glmatrix.vec3.fromValues(this.linearScalingFactor,
-              this.linearScalingFactor,
-              this.linearScalingFactor)
-
-          // Scale the matrix
-          glmatrix.mat4.scale(scaleMatrix, scaleMatrix, scaleVec)
-
-          // Perform the matrix multiplications
-          if (newMatrix !== void 0) {
-            glmatrix.mat4.multiply(newTransform, coordinationMatrix, newMatrix)
-            glmatrix.mat4.multiply(newTransform, newTransform, translationMatrixGeomMin)
-          } else {
-            glmatrix.mat4.multiply(newTransform, coordinationMatrix, newTransform)
-            glmatrix.mat4.multiply(newTransform, newTransform, translationMatrixGeomMin)
-          }
-          const newTransformArr = Array.from(newTransform)
-          geometryMaterialTransformMap.set(expressID,
-              [geometry.geometry, material_, newTransformArr])
-
-          if (entity?.localID !== void 0) {
-            if (entity?.expressID !== void 0) {
-              const mesh = meshMap.get(entity.localID)
-              if (mesh !== void 0) {
-                // set color
-                const color = {
-                  x: material_.legacyColor[0],
-                  y: material_.legacyColor[1],
-                  z: material_.legacyColor[2],
-                  w: material_.legacyColor[3],
-                }
-
-                // Single PlacedGeometry variable
-                const singlePlacedGeometry: PlacedGeometry = {
-                  color: color,
-                  geometryExpressID: expressID,
-                  flatTransformation: newTransformArr,
-                }
-
-                mesh[0].push(singlePlacedGeometry)
-                mesh[1].geometries = mesh[0]
-
-                meshMap.set(entity.localID, [mesh[0], mesh[1]])
-
-
-              } else {
-                // set color
-                const color = {
-                  x: material_.legacyColor[0],
-                  y: material_.legacyColor[1],
-                  z: material_.legacyColor[2],
-                  w: material_.legacyColor[3],
-                }
-
-                // Single PlacedGeometry variable
-                const singlePlacedGeometry_: PlacedGeometry = {
-                  color: color,
-                  geometryExpressID: expressID,
-                  flatTransformation: newTransformArr,
-                }
-
-                // eslint-disable-next-line no-array-constructor
-                const placedGeometryArray_ = new Array<PlacedGeometry>()
-
-                // Vector of PlacedGeometry
-                const vectorOfPlacedGeometry_: Vector<PlacedGeometry> = {
-                  get(index: number): PlacedGeometry {
-                    if (index >= placedGeometryArray_.length) {
-                      return singlePlacedGeometry_
-                    }
-
-                    return placedGeometryArray_[index]
-                  },
-                  size(): number {
-                    return placedGeometryArray_.length
-                  },
-                  push(parameter: PlacedGeometry): void {
-                    placedGeometryArray_.push(parameter)
-                  },
-                }
-
-                vectorOfPlacedGeometry_.push(singlePlacedGeometry_)
-
-                const singleFlatMesh: FlatMesh = {
-                  geometries: vectorOfPlacedGeometry_,
-                  expressID: entity.expressID,
-                }
-
-                meshMap.set(entity.localID, [vectorOfPlacedGeometry_, singleFlatMesh])
-              }
-            }
-          }
-        }
-      }
-
-
-      meshMap.forEach((mesh, productLocalID) => {
-
-        vectorFlatMesh.push(mesh[1])
-
-        meshCallback(mesh[1])
-      })
+      result.streamAllMeshesWithTypes(types, meshCallback)
     }
   }
 
@@ -1327,225 +493,8 @@ export class IfcAPI {
     const result = this.models.get(modelID)
 
     if (result !== void 0) {
-      const [model,
-        scene,
-        meshMap,
-        geometryMaterialTransformMap,
-        vectorFlatMesh] = result
 
-      let coordinationMatrix = result[5]
-
-      // eslint-disable-next-line no-unused-vars
-      for (const [_, nativeTransform, geometry, material, entity] of scene.walk()) {
-
-        if (geometry.type === CanonicalMeshType.BUFFER_GEOMETRY && !geometry.temporary) {
-          let material_: CanonicalMaterial | undefined
-          if (material === void 0) {
-            material_ = {
-              name: '',
-              // eslint-disable-next-line no-magic-numbers
-              baseColor: [0.8, 0.8, 0.8, 1],
-              // eslint-disable-next-line no-magic-numbers
-              legacyColor: [0.8, 0.8, 0.8, 1],
-              doubleSided: true,
-              blend: BlendMode.OPAQUE,
-            }
-          } else {
-            material_ = material
-          }
-
-          // extract min
-          const geomMin: glmatrix.vec3 = glmatrix.vec3.create()
-          const minPt = geometry.geometry.getMin()
-          geomMin[0] = minPt.x
-          geomMin[1] = minPt.y
-          geomMin[2] = minPt.z
-
-          // Create a translation matrix from geom.min
-          const translationMatrixGeomMin: glmatrix.mat4 = glmatrix.mat4.create()
-          glmatrix.mat4.fromTranslation(translationMatrixGeomMin, geomMin)
-
-          // create PlacedGeometry
-          const expressID = model.getElementByLocalID(geometry.localID)?.expressID as number
-
-          const geometryTransform = nativeTransform?.getValues()
-          let newMatrix: glmatrix.mat4 | undefined
-          if (geometryTransform !== void 0) {
-            newMatrix = glmatrix.mat4.fromValues(
-                geometryTransform[0],
-                geometryTransform[1],
-                geometryTransform[2],
-                geometryTransform[3],
-                geometryTransform[4],
-                geometryTransform[5],
-                geometryTransform[6],
-                geometryTransform[7],
-                geometryTransform[8],
-                geometryTransform[9],
-                geometryTransform[10],
-                geometryTransform[11],
-                geometryTransform[12],
-                geometryTransform[13],
-                geometryTransform[14],
-                geometryTransform[15],
-            )
-          }
-
-          if (!this._isCoordinated && this.settings?.COORDINATE_TO_ORIGIN) {
-            // coordinate the geometry to the origin
-            // Assuming geom.GetPoint(0) returns a glm::dvec3, i.e., a 3D vector.
-            // In TypeScript, you can represent it as number[] or Float64Array.
-            Logger.info('Setting up coordinationMatrix')
-
-            // eslint-disable-next-line new-cap
-            const nativePt = geometry.geometry.GetPoint(0)
-            const pt: number[] = [nativePt.x, nativePt.y, nativePt.z]
-
-            // Transform the point by the matrix.
-            const transformedPt: glmatrix.vec4 = glmatrix.vec4.create()
-            glmatrix.vec4.transformMat4(transformedPt, [pt[0], pt[1], pt[2], 1], newMatrix!)
-
-            // Create the translation matrix.
-            coordinationMatrix = glmatrix.mat4.create()
-
-            glmatrix.mat4.fromTranslation(coordinationMatrix,
-                [-transformedPt[0], -transformedPt[1], -transformedPt[2]])
-
-            const scaleMatrix = glmatrix.mat4.create()
-
-            // Create a 3D vector for scaling factors
-            const scaleVec = glmatrix.vec3.fromValues(this.linearScalingFactor,
-                this.linearScalingFactor,
-                this.linearScalingFactor)
-
-            // Scale the matrix
-            glmatrix.mat4.scale(scaleMatrix, scaleMatrix, scaleVec)
-
-            glmatrix.mat4.multiply(coordinationMatrix,
-                this.NormalizeMat,
-                coordinationMatrix)
-            glmatrix.mat4.multiply(coordinationMatrix,
-                scaleMatrix,
-                coordinationMatrix)
-
-            this._isCoordinated = true
-          }
-
-          // normalize geometry
-          if (!geometry.geometry.normalized) {
-            // eslint-disable-next-line new-cap
-            geometry.geometry.NormalizeInPlace()
-          }
-
-
-          // extract color
-          const newTransform = glmatrix.mat4.create()
-
-          // Create a 4x4 identity matrix
-          const scaleMatrix = glmatrix.mat4.create()
-
-          // Create a 3D vector for scaling factors
-          const scaleVec = glmatrix.vec3.fromValues(this.linearScalingFactor,
-              this.linearScalingFactor,
-              this.linearScalingFactor)
-
-          // Scale the matrix
-          glmatrix.mat4.scale(scaleMatrix, scaleMatrix, scaleVec)
-
-          // Perform the matrix multiplications
-          if (newMatrix !== void 0) {
-            glmatrix.mat4.multiply(newTransform, coordinationMatrix, newMatrix)
-            glmatrix.mat4.multiply(newTransform, newTransform, translationMatrixGeomMin)
-          } else {
-            glmatrix.mat4.multiply(newTransform, coordinationMatrix, newTransform)
-            glmatrix.mat4.multiply(newTransform, newTransform, translationMatrixGeomMin)
-          }
-          const newTransformArr = Array.from(newTransform)
-          geometryMaterialTransformMap.set(expressID,
-              [geometry.geometry, material_, newTransformArr])
-
-          if (entity?.localID !== void 0) {
-            if (entity?.expressID !== void 0) {
-              const mesh = meshMap.get(entity.localID)
-              if (mesh !== void 0) {
-                // set color
-                const color = {
-                  x: material_.legacyColor[0],
-                  y: material_.legacyColor[1],
-                  z: material_.legacyColor[2],
-                  w: material_.legacyColor[3],
-                }
-
-                // Single PlacedGeometry variable
-                const singlePlacedGeometry: PlacedGeometry = {
-                  color: color,
-                  geometryExpressID: expressID,
-                  flatTransformation: newTransformArr,
-                }
-
-                mesh[0].push(singlePlacedGeometry)
-                mesh[1].geometries = mesh[0]
-
-                meshMap.set(entity.localID, [mesh[0], mesh[1]])
-
-
-              } else {
-                // set color
-                const color = {
-                  x: material_.legacyColor[0],
-                  y: material_.legacyColor[1],
-                  z: material_.legacyColor[2],
-                  w: material_.legacyColor[3],
-                }
-
-                // Single PlacedGeometry variable
-                const singlePlacedGeometry_: PlacedGeometry = {
-                  color: color,
-                  geometryExpressID: expressID,
-                  flatTransformation: newTransformArr,
-                }
-
-                // eslint-disable-next-line no-array-constructor
-                const placedGeometryArray_ = new Array<PlacedGeometry>()
-
-                // Vector of PlacedGeometry
-                const vectorOfPlacedGeometry_: Vector<PlacedGeometry> = {
-                  get(index: number): PlacedGeometry {
-                    if (index >= placedGeometryArray_.length) {
-                      return singlePlacedGeometry_
-                    }
-
-                    return placedGeometryArray_[index]
-                  },
-                  size(): number {
-                    return placedGeometryArray_.length
-                  },
-                  push(parameter: PlacedGeometry): void {
-                    placedGeometryArray_.push(parameter)
-                  },
-                }
-
-                vectorOfPlacedGeometry_.push(singlePlacedGeometry_)
-
-                const singleFlatMesh: FlatMesh = {
-                  geometries: vectorOfPlacedGeometry_,
-                  expressID: entity.expressID,
-                }
-
-                meshMap.set(entity.localID, [vectorOfPlacedGeometry_, singleFlatMesh])
-              }
-            }
-          }
-        }
-      }
-
-
-      meshMap.forEach((mesh, productLocalID) => {
-
-        vectorFlatMesh.push(mesh[1])
-      })
-
-      return vectorFlatMesh
+      return result.loadAllGeometry()
     }
 
     // dummy vars
@@ -1621,18 +570,8 @@ export class IfcAPI {
     const result = this.models.get(modelID)
 
     if (result !== void 0) {
-      // eslint-disable-next-line no-unused-vars
-      const [model, scene, meshMap] = result
-      if (meshMap.size <= 0) {
-        // eslint-disable-next-line new-cap
-        this.LoadAllGeometry(modelID)
-      }
 
-      const mesh = meshMap.get(expressID)
-
-      if (mesh !== void 0) {
-        return mesh[1]
-      }
+      return result.getFlatMesh(expressID)
     }
 
     // Single PlacedGeometry variable
